// backend/src/catalog/types/questionSetTypes.ts
export type InputType = keyof typeof InputTypeEnum;

// Base Types
export type TQuestionSetBase = {
  internalName: string;
  displayName: string;
  description?: string;
  client_id: number;
};

export type TQuestionBase = {
  key?: string;
  title: string;
  inputType: InputType;
  tooltip?: string;
  category?: string;
  isRequired: boolean;
  orderNo: number;
  metadata?: Record<string, any>;
};

export type TQuestionOptionBase = {
  key?: string;
  title: string;
  priceModifier?: number;
  icon?: string;
  isDefault: boolean;
  orderNo: number;
  metadata?: Record<string, any>;
};

export type TQuestionSetAssignmentBase = {
  modelId: number;
  questionSetId: number;
  assignmentOrder: number;
};

// Create Types
export type TQuestionSetCreate = TQuestionSetBase;
export type TQuestionCreate = TQuestionBase & { questionSetId: number };
export type TQuestionOptionCreate = TQuestionOptionBase & { questionId: number };
export type TQuestionSetAssignmentCreate = TQuestionSetAssignmentBase;

// Translation Create Types
export type TQuestionSetTranslationCreate = {
  questionSetId: number;
  languageId: number;
  displayName: string;
  description?: string;
};

export type TQuestionTranslationCreate = {
  questionId: number;
  languageId: number;
  title: string;
  tooltip?: string;
  category?: string;
};

export type TQuestionOptionTranslationCreate = {
  optionId: number;
  languageId: number;
  title: string;
};

// Update Types
export type TQuestionSetUpdate = Partial<TQuestionSetBase>;
export type TQuestionUpdate = Partial<TQuestionBase>;
export type TQuestionOptionUpdate = Partial<TQuestionOptionBase>;